% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_fetch.R
\name{fs_is_absolute}
\alias{fs_is_absolute}
\title{Is a path absolute? (handles POSIX and Windows drive-letter/UNC paths)}
\usage{
fs_is_absolute(path)
}
\description{
Is a path absolute? (handles POSIX and Windows drive-letter/UNC paths)
}
\keyword{internal}
