% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_branch_cache_read}
\alias{tf_branch_cache_read}
\title{Read the branch-list cache from disk, if present}
\usage{
tf_branch_cache_read()
}
\description{
Read the branch-list cache from disk, if present
}
\keyword{internal}
