% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_branch_cache_write}
\alias{tf_branch_cache_write}
\title{Write the branch-list cache to disk}
\usage{
tf_branch_cache_write(branches)
}
\description{
Write the branch-list cache to disk
}
\keyword{internal}
