% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_cache_read}
\alias{tf_cache_read}
\title{Read the folder-list cache for a branch from disk, if present}
\usage{
tf_cache_read(branch = NULL)
}
\description{
Read the folder-list cache for a branch from disk, if present
}
\keyword{internal}
