% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_cache_stale}
\alias{tf_cache_stale}
\title{Is the cache stale (older than max_age_days, or missing)?}
\usage{
tf_cache_stale(cache, max_age_days = 7)
}
\description{
Is the cache stale (older than max_age_days, or missing)?
}
\keyword{internal}
