% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_cache_write}
\alias{tf_cache_write}
\title{Write the folder-list cache for a branch to disk}
\usage{
tf_cache_write(folders, branch = NULL)
}
\description{
Write the folder-list cache for a branch to disk
}
\keyword{internal}
