% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_fetch.R
\name{tf_download_folder}
\alias{tf_download_folder}
\title{Download a folder recursively via the GitHub contents API}
\usage{
tf_download_folder(folder_name, dest_dir, quiet = FALSE)
}
\description{
Download a folder recursively via the GitHub contents API
}
\keyword{internal}
