% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_fetch_branches_remote}
\alias{tf_fetch_branches_remote}
\title{Query the GitHub API for branches in the repo}
\usage{
tf_fetch_branches_remote()
}
\description{
Query the GitHub API for branches in the repo
}
\keyword{internal}
