% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_fetch.R
\name{tf_install_package}
\alias{tf_install_package}
\title{Install a downloaded folder as an R package via pak}
\usage{
tf_install_package(dest_dir, quiet = FALSE)
}
\description{
Install a downloaded folder as an R package via pak
}
\keyword{internal}
