% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_fetch_folders_remote}
\alias{tf_fetch_folders_remote}
\title{Query the GitHub contents API for top-level folders in the repo}
\usage{
tf_fetch_folders_remote()
}
\description{
Query the GitHub contents API for top-level folders in the repo
}
\keyword{internal}
