% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_get_folders}
\alias{tf_get_folders}
\title{Get the list of top-level folders, using cache unless stale or forced}
\usage{
tf_get_folders(force = FALSE, max_age_days = 7, quiet = FALSE)
}
\description{
Get the list of top-level folders, using cache unless stale or forced
}
\keyword{internal}
